requirements.txt
anyio==4.8.0
asttokens==3.0.0
attrs==24.3.0
Automat==24.8.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
constantly==23.10.4
cryptography==44.0.0
cssselect==1.2.0
decorator==5.1.1
defusedxml==0.7.1
executing==2.1.0
filelock==3.16.1
Flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipython==8.31.0
itemadapter==0.10.0
itemloaders==1.3.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mysql-connector-python==9.1.0
packaging==24.2
parsel==1.9.1
parso==0.8.4
pexpect==4.9.0
praw==7.8.1
prawcore==2.4.0
prompt_toolkit==3.0.48
Protego==0.3.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
Pygments==2.19.1
PyMySQL==1.1.1
pyOpenSSL==25.0.0
python-dotenv==1.0.1
python-telegram-bot==21.10
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
Scrapy==2.12.0
service-identity==24.2.0
setuptools==75.8.0
sniffio==1.3.1
stack-data==0.6.3
tldextract==5.1.3
traitlets==5.14.3
Twisted==24.11.0
typing_extensions==4.12.2
update-checker==0.18.0
urllib3==2.3.0
w3lib==2.2.1
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
zope.interface==7.2
